+++ packages/terra-worklist-data-grid/CHANGELOG.md
@@ -1,15 +1,13 @@
 # Changelog
 
 ## Unreleased
-* Added
-  * Added masked cell support.
-
 * Changed
   * Removed ability to set variable row heights
 * Added
   * Added masked cell support.
   * Removed ability to set variable row heights.
   * Updated styling on cell focus.
+  * Added hover state for selectable cells.
 
 ## 0.1
   * Initial stable release.